+++ src/schema/opinionNewsArticle/index.ts
@@ -1,0 +1,3 @@
+export {buildOpinionNewsArticleJsonLd, OpinionNewsArticleSchema} from './component'
+export {opinionNewsArticleFields, default as schemaOrgOpinionNewsArticle} from './schema'
+export type {SchemaOrgOpinionNewsArticleConfig, SchemaOrgOpinionNewsArticleData} from './types'